Split refresh into coordination and execution layers
